% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizing_betas.R
\name{betas}
\alias{betas}
\title{Finding the Betas}
\usage{
betas(Beta, f, x, y, Norm)
}
\arguments{
\item{Beta}{A \code{vector} containing the desired starting coeffs for estimation}

\item{f}{A \code{function} that we are optimizing. Default is \code{least_square} from the same package}

\item{x}{A \code{matrix} containing predictors}

\item{y}{A \code{vector} containing outputs}

\item{Norm}{A \code{character} defining loss to use (either 'L1' or 'L2')}
}
\value{
A \code{vector} giving the optimized beta values
}
\description{
This function will spit out the desired \code{beta} values for the
least squares function
}
\examples{
  x <- matrix(rnorm(600), nrow=200)
  epsilon <- rnorm(200, 0, sd=.25)
  x <- cbind(rep(1,200), x)
  actual_beta <- rnorm(4,5,10)
  y <- x\%*\%actual_beta + epsilon
  start_beta <- rnorm(4)
  betas(Beta=start_beta,y=y, x=x, Norm="L2")
}
\author{
Charles Benfer
}
