% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{squared_error}
\alias{squared_error}
\title{Evaluate Linear Loss Function}
\usage{
squared_error(Beta, Y, X, norm)
}
\arguments{
\item{Beta}{A \code{vector} containing coeffs}

\item{Y}{A \code{vector} containing outputs}

\item{X}{A \code{matrix} containing predictors}

\item{norm}{A \code{character} defining loss to use (default is 'L2')}
}
\value{
A \code{numeric} giving value of loss at \code{beta}
}
\description{
This function delivers the value of the loss function for a given
value of \code{beta}
}
\examples{
  x <- matrix(rnorm(600), nrow=200)
  epsilon <- rnorm(200, 0, sd=.25)
  x <- cbind(rep(1,200), x)
  actual_beta <- rnorm(4,5,10)
  y <- x\%*\%actual_beta + epsilon
  start_beta <- rnorm(4)
  squared_error(start_beta, Y=y, X=x, norm="L2")
}
\author{
Charles Benfer
}
